Create new imputation module with imputation codes table
