+++ src/pudl/analysis/__init__.py
@@ -10,6 +10,7 @@ from . import (
     allocate_gen_fuel,
     epacamd_eia,
     fuel_by_plant,
+    imputation,
     mcoe,
     plant_parts_eia,
     pudl_models,
@@ -17,5 +18,4 @@ from . import (
     service_territory,
     spatial,
     state_demand,
-    timeseries_cleaning,
 )
