+++ src/pudl/etl/__init__.py
@@ -69,7 +69,7 @@ core_module_groups = {
         pudl.convert.censusdp1tract_to_sqlite,
         pudl.output.censusdp1tract,
     ],
-    "core_codes": [static_assets],
+    "core_codes": [static_assets, pudl.analysis.imputation.codes],
     "core_eia": [pudl.transform.eia],
     "core_eiaaeo": [pudl.transform.eiaaeo],
     "core_eia_bulk_elec": [eia_bulk_elec_assets],
@@ -95,7 +95,6 @@ out_module_groups = {
     "out_eia": [
         pudl.output.eia,
         pudl.output.eia860,
-        pudl.output.eia930,
         pudl.output.eia923,
         pudl.output.eia_bulk_elec,
     ],
