+++ src/pudl/metadata/resources/pudl.py
@@ -32,6 +32,24 @@ RESOURCE_METADATA: dict[str, dict[str, Any]] = {
         "field_namespace": "pudl",
         "sources": ["pudl"],
     },
+    "core_pudl__codes_imputation_reasons": {
+        "title": "Imputation Reason Codes",
+        "description": (
+            "Contains codes and descriptions, which indicate why a value was flagged."
+            "for imputation. Any time a column is imputed, there will be a corresponding"
+            "column which contains codes for every value in the column that was imputed."
+        ),
+        "schema": {
+            "fields": [
+                "code",
+                "description",
+            ],
+            "primary_key": ["code"],
+        },
+        "etl_group": "static_pudl",
+        "field_namespace": "pudl",
+        "sources": ["pudl"],
+    },
     "core_pudl__entity_plants_pudl": {
         "title": "PUDL Plants",
         "description": "Home table for PUDL assigned plant IDs. These IDs are manually generated each year when new FERC and EIA reporting is integrated, and any newly identified plants are added to the list with a new ID. Each ID maps to a power plant which is reported in at least one FERC or EIA data set. This table is read in from a spreadsheet stored in the PUDL repository: src/pudl/package_data/glue/pudl_id_mapping.xlsx",
