+++ src/pudl/output/__init__.py
@@ -16,7 +16,6 @@ from . import (
     eia,
     eia860,
     eia923,
-    eia930,
     eia_bulk_elec,
     epacems,
     ferc1,
